Remove the non-colon ERT interword spaces
Remove the ERTs that have "\ ", which is an interword space, if they
don't follow a colon.

I checked with diffpdf that the ones remove in this commit do not
have an effect on output. e.g., they do not have an effect on output
if following "OLS" or "IV" or other capital letters. They would have
an effect if they followed a period, or a colon.

This is the command I used:

sed -z -i 's/\([^:]\)\n\\begin_inset ERT\nstatus collapsed\n\n\\begin_layout Plain Layout\n\n\n\\backslash\n \n\\end_layout\n\n\\end_inset/\1 /g' econometrics.lyx

I had to correct one special case that the above sed removed, and
diffpdf caught, which was due to a colon in bold font (the above
regex was tricked because the bold font ended after the colon).