Document settings: set all math packages to auto
In previous versions of LyX, it was necessary to work around some
problems, either by always loading a math package (because LyX did
not detect it needed to be loaded) or by never loading it (because
LyX incorrectly loaded it). Now, those problems have been ironed
out.

diffpdf confirms the output is equivalent.